+++ app/Http/Controllers/CategoryController.php
@@ -11,6 +11,10 @@ class CategoryController extends Controller
 {
     public function index() 
     {
+        if (!auth()->user()->can('category-list')) {
+            abort(403, 'Access Denied');
+        }
+
         return view('category.index');
     }
 
@@ -20,16 +24,22 @@ class CategoryController extends Controller
         $datatables = DataTables::of($data)
         ->addIndexColumn()
         ->addColumn('action', function($row){
-            $btn = '<a href ="'.route('category.edit', $row->slug).'" class="btn btn-primary btn-sm me-2 mb-2">
-                        <i class="fas fa-edit"></i>
-                    </a>';
-            $btn .= '<form id="delete-category-'.$row->id.'" action="'.route('category.destroy', $row->slug).'" method="POST" style="display:inline;">'
-                    .csrf_field()
-                    .method_field('DELETE')
-                    .'<button type="button" class="btn btn-danger btn-sm mb-2" onclick="confirmDelete(\'delete-category-'.$row->id.'\')">
-                        <i class="fas fa-trash"></i>
-                    </button>'
-                .'</form>';
+            $btn = '';
+
+            if (auth()->user()->can('category-edit')) {
+                $btn = '<a href ="'.route('category.edit', $row->slug).'" class="btn btn-primary btn-sm me-2 mb-2">
+                            <i class="fas fa-edit"></i>
+                        </a>';
+            }
+            if (auth()->user()->can('category-delete')) {
+                $btn .= '<form id="delete-category-'.$row->id.'" action="'.route('category.destroy', $row->slug).'" method="POST" style="display:inline;">'
+                        .csrf_field()
+                        .method_field('DELETE')
+                        .'<button type="button" class="btn btn-danger btn-sm mb-2" onclick="confirmDelete(\'delete-category-'.$row->id.'\')">
+                            <i class="fas fa-trash"></i>
+                        </button>'
+                    .'</form>';
+            }
 
             return $btn;
         })
@@ -41,11 +51,17 @@ class CategoryController extends Controller
 
     public function create()
     {
+        if (!auth()->user()->can('category-create')) {
+            abort(403, 'Access Denied');
+        }
         return view('category.create');
     }
     
     public function store(Request $request)
     {
+        if (!auth()->user()->can('category-create')) {
+            abort(403, 'Access Denied');
+        }
         $request->validate([
             'title' => 'required',
         ]);
@@ -55,17 +71,23 @@ class CategoryController extends Controller
         $category->slug = $this->generateUniqueSlug($request->title);
         $category->save();
 
-        return redirect()->route('category')->with('success', 'Category berhasil ditambahkan');
+        return redirect()->route('category')->with('success', 'Category Successfully Added');
     }
 
     public function edit($slug)
     {
+        if (!auth()->user()->can('category-edit')) {
+            abort(403, 'Access Denied');
+        }
         $category = Category::where('slug', $slug)->first();
         return view('category.edit', compact('category'));
     }
     
     public function update(Request $request,  $slug)
     {
+        if (!auth()->user()->can('category-edit')) {
+            abort(403, 'Access Denied');
+        }
         $request->validate([
             'title' => 'required',
         ]);
@@ -75,15 +97,19 @@ class CategoryController extends Controller
         $category->slug = $this->generateUniqueSlug($request->title, $category->id);
         $category->save();
 
-        return redirect()->route('category')->with('success', 'Category berhasil diperbarui');
+        return redirect()->route('category')->with('success', 'Category Successfully Updated');
     }
 
     public function destroy($slug)
     {
+        if (!auth()->user()->can('category-delete')) {
+            abort(403, 'Access Denied');
+        }
+
         $category = Category::where('slug', $slug)->first();
         $category->delete();
 
-        return redirect()->back()->with('success', 'Category berhasil dihapus');
+        return redirect()->back()->with('success', 'Category Successfully Deleted');
     }
 
     // Helper untuk generate slug unik
